refactor(pcap): extract fixture generation into standalone example
Move generate_fixtures from an ignored test in pcap.rs to a standalone
example gated behind the pcap-replay feature flag. Make PcapPacket,
parse_file, and write_file public so the example can use them. Add a
Makefile `fixtures` target that rebuilds fixtures automatically.
