+++ Cargo.toml
@@ -115,3 +115,7 @@ reqwest = { version = "0.12.28", features = ["json"] }
 tokio-test = "0.4"
 tempfile = "3"
 
+[[example]]
+name = "generate-fixtures"
+required-features = ["pcap-replay"]
+
